LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-001-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 2 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 3 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
